requirements.txt
Flask
python-bitcoinrpc
